Test: LottoAmount에 대한 단위 테스트 코드 구현
